GW-8: Add Extensions and Dao Tests

+++ app/src/main/java/com/greyp/weather/data/local/entities/embedded/RainEntity.kt
@@ -9,7 +9,7 @@ import com.greyp.weather.utils.Constants
 data class RainEntity(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = "rain_volume_in_1h")
-    val volumeIn1h: Int,
+    val volumeIn1h: Double,
     @ColumnInfo(name = "rain_volume_in_3h")
-    val volumeIn3h: Int
+    val volumeIn3h: Double
 )
