+++ app/src/main/java/com/greyp/weather/data/local/entities/embedded/SnowEntity.kt
@@ -9,7 +9,7 @@ import com.greyp.weather.utils.Constants
 data class SnowEntity(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = "snow_volume_in_1h")
-    val volumeIn1h: Int,
+    val volumeIn1h: Double,
     @ColumnInfo(name = "snow_volume_in_3h")
-    val volumeIn3h: Int
+    val volumeIn3h: Double
 )
